% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_handling_for_unit.R
\name{build_handling_codes}
\alias{build_handling_codes}
\title{Build Special Handling Codes}
\usage{
build_handling_codes(
  no_contact,
  no_solc,
  no_mail,
  no_email,
  no_phone,
  no_sms,
  other_codes
)
}
\arguments{
\item{no_contact}{a logical}

\item{no_solc}{a logical}

\item{no_mail}{a logical}

\item{no_email}{a logical}

\item{no_phone}{a logical}

\item{no_sms}{a logical}

\item{other_codes}{a character vector of any codes to add manually}
}
\value{
a vector of unique special handling codes based on the logical flags specified
}
\description{
Build Special Handling Codes
}
