% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReconnect.R
\name{dbReconnect}
\alias{dbReconnect}
\title{dbReconnect}
\usage{
dbReconnect(
  con = cdw,
  reconnect_tables = TRUE,
  reconnect_all_tbls_with_this_con = FALSE
)
}
\arguments{
\item{con}{a existing database connection to reconnect to}

\item{reconnect_tables}{a logical value specifying whether all database tables should be reconncted using con. If you have more than one database connection, you should set this to FALSE, as dbReconnect() has not way to know which connection you used to create a table and will try to use con to recreate all your tbl_sql object with con.}

\item{reconnect_all_tbls_with_this_con}{a logical value used to override warnings about multiple database connections. FALSE by default, meaning this function will error out if you have multiple database connections and try to reconnect tables. Set to TRUE to override the warning and reconnect all tables with the connection specified in con.}
}
\value{
returns TRUE invisibly
}
\description{
reconnect to a database and reinitialize all the tables. NOTE: doesn't work right if you have more than one active database connection.
}
\details{
This function writes to the global environment, overwriting an existing database connection with a valid one, and if reocnnect_tables = TRUE, reconnecting all tbl_sql objects with the newly written, valid connection. Note that this can be really problematic if you have more than one active database connection--there's not a good way to know which connetion was used for each table, so the function will error out if you have more than one database connection and are trying to reconnec all the tables.
}
