% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReconnect.R
\name{dbRemakeTbl}
\alias{dbRemakeTbl}
\title{Reconnect a database table}
\usage{
dbRemakeTbl(thetbl, con = cdw, assign_global = T)
}
\arguments{
\item{thetbl}{a tbl_sql to reconnect, or a character string of the object name. Note that when thetbl is specified as a string, and the connection has already been disconnected, the function cannot reconnect.}

\item{con}{a database connection to use}

\item{assign_global}{a logical value determining whether or not the reconnected table should be written invisibly back to the global environment or should be returned.}
}
\value{
if assign_global is TRUE, then returns TRUE invisibly. Otherwise, it returns the reconnected table
}
\description{
Reconnect a database table
}
