% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_git_commits.R
\name{get_git_commits}
\alias{get_git_commits}
\title{Get Commit Data from Gitlab}
\usage{
get_git_commits(
  project_id = 102,
  token = Sys.getenv("GITLAB_TOKEN"),
  return = "tableau",
  timezone = "America/Chicago"
)
}
\arguments{
\item{project_id}{the id of a project. See project/settings/general}

\item{token}{a private token, stored in .Renviron by default. Create a token in User Settings/Access Tokens}

\item{return}{how you want the data returned. Choose "tableau" for a string vector you can paste in Tableau, or "string" for an unformatted character vector. Choose "data_frame" for a table of values. Choose "list" for a list with the table as one item and the string as the second item.}
}
\value{
up to 10000 commits from gitlab. How this is returned depends on the value of return
}
\description{
Get Commit Data from Gitlab
}
