% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_handling_for_unit.R
\name{get_handling_for_unit}
\alias{get_handling_for_unit}
\title{Get Special Handling Codes for A Unit}
\usage{
get_handling_for_unit(
  no_contact,
  no_solc,
  no_mail,
  no_email,
  no_phone,
  no_sms,
  other_codes,
  unit_codes = c("UC"),
  group_codes = NA,
  handling_tbl = tbl(cdw, in_schema2("CDW", "d_bio_handling_mv"))
)
}
\arguments{
\item{no_contact}{a logical}

\item{no_solc}{a logical}

\item{no_mail}{a logical}

\item{no_email}{a logical}

\item{no_phone}{a logical}

\item{no_sms}{a logical}

\item{other_codes}{a character vector of any codes to include. Use this if you don't want to use the logical flags}

\item{unit_codes}{a character vector of unit codes to include. 'UC' is included by default and added to all queries if you don't pass it in}

\item{group_codes}{a character vector of group codes to include. If NA, rows with blank group codes are returned}

\item{handling_tbl}{a tbl_Oracle where the handling data lives}
}
\value{
a tbl_Oracle with one row per unique entity_id and handling type code pair
}
\description{
Get Special Handling Codes for A Unit
}
