% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dictionary.R
\name{get_main_dictionary}
\alias{get_main_dictionary}
\title{Get Data Dictionary}
\usage{
get_main_dictionary(cdw_only = T)
}
\arguments{
\item{cdw_only}{a logical determining whether we should just grab the CDW-related columns (and DATA_DESCRIPTION)}
}
\value{
a data_frame of the main big data dictionary
}
\description{
Get Data Dictionary
}
