% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualification_date.R
\name{get_qualification_date}
\alias{get_qualification_date}
\title{Get Qualification DAte}
\usage{
get_qualification_date(db = cdw)
}
\arguments{
\item{db}{a valid database connection}
}
\value{
a database tibble with one line per proposal and the qualification date calculated in the PROSPECT data sources.
}
\description{
Get Qualification DAte
}
