% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSQL.R
\name{getSQL}
\alias{getSQL}
\title{get SQL from file}
\usage{
getSQL(filepath, returnsql = T)
}
\arguments{
\item{filepath}{the path of an SQL script.}

\item{returnsql}{a logical value determining whether the function should return an SQL object (via dplyr::sql()) or just a text string}
}
\value{
either an sql object or a text string to be used in a tbl() connection
}
\description{
get SQL from file
}
