% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_schema2.R
\name{in_schema2}
\alias{in_schema2}
\title{In Schema2}
\usage{
in_schema2(schema, table)
}
\arguments{
\item{schema}{a database schema name}

\item{table}{the name of a database table}
}
\description{
a convenience wrapper around \link[dbplyr]{in_schema} that capitalizes the schema and table names. Apparently, some combination of recent versions of ODBC and/or R are case-sensitive with regard to table names (this was not the case prior to October 2020). So, this just capitalizes the table names so you don't have to fight it.
}
