% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all.R
\name{print_all}
\alias{print_all}
\title{Print All}
\usage{
print_all(data, n = 1000, ...)
}
\arguments{
\item{data}{an object to print to the console}

\item{n}{number of rows to print. Defaults to 1000, unlike print.tibble, which prints 10-15 rows by default}

\item{...}{other options passed to print()}
}
\value{
returns the results of print(data), usually data itself
}
\description{
a quick helper function to print all the rows of a tibble to the console. Basically a wrapper for print(data, n = 1000)
}
