% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dictionary.R
\name{search_dictionary}
\alias{search_dictionary}
\title{Search the Data Dictionary}
\usage{
search_dictionary(
  search_term,
  cdw_table,
  cdw_col,
  the_dictionary = get_main_dictionary()
)
}
\arguments{
\item{search_term}{a regular expression to search across for within every column}

\item{cdw_table}{a regular expression to search for in the CDW table name column}

\item{cdw_col}{a regular expression to search for in the CDW column name column}

\item{the_dictionary}{a dictionary object to search within, as created by get_main_dictionary(). By default, pulls the CDW-related columns.}
}
\value{
a filtered data_Frame from the_dictionary matching the specified search terms
}
\description{
Search the Data Dictionary
}
